# chore(Spanner): Make createTransactionSelector faster