--- a/Spanner/src/Connection/Grpc.php
+++ b/Spanner/src/Connection/Grpc.php
@@ -1387,9 +1387,10 @@ class Grpc implements ConnectionInterface
                 $transaction['begin'] = $this->formatTransactionOptions($transaction['begin']);
             }
 
-            $selector = $this->serializer->decodeMessage($selector, $transaction);
+            $str = json_encode($transaction);
+            $selector->mergeFromJsonString($str);
         } elseif (isset($args['transactionId'])) {
-            $selector = $this->serializer->decodeMessage($selector, ['id' => $this->pluck('transactionId', $args)]);
+            $selector->setId($this->pluck('transactionId', $args));
         }
 
         return $selector;