@saranshdhingra 2024-05-30
Seems like protobuf's `mergeFromString` doesn't work with the following:
`{"begin": {"readWrite": []}}` whereas it works with: `{"begin": {"readWrite": {}}}`.

For now, closing this PR.