@bshaffer 2024-06-13
> Seems like protobuf's mergeFromString doesn't work with the following:
{"begin": {"readWrite": []}} whereas it works with: {"begin": {"readWrite": {}}}.

I have made sure that the JSON string uses `{}` for empty arrays by providing the `JSON_FORCE_OBJECT` flag, but the emulator tests are still failing. This will require further investigation.