@bshaffer 2025-12-23
Closing as Spanner has been refactored and this is no longer relevant